requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
tensorflow>=2.8.0
pillow>=8.0.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
pandas>=1.3.0
streamlit>=1.0.0